move deserialized_column to its own file
